refactor: `EnvironmentBanner`
