+++ App.tsx
@@ -7,6 +7,7 @@ import { getEnvironment } from "./constants/settings";
 import useCachedResources from "./hooks/useCachedResources";
 import Navigation from "./navigation";
 import store from "./store/config";
+import { EnvironmentBanner } from "./utils/EnvironmentBanner";
 
 export default function App() {
   const isLoadingComplete = useCachedResources();
@@ -22,6 +23,7 @@ export default function App() {
             environment={getEnvironment()}
             languageCode="no"
           />
+          <EnvironmentBanner />
           <Navigation colorScheme="light" />
           <StatusBar style="dark" />
         </SafeAreaProvider>
