+++ navigation/utils.js
@@ -1,11 +1,5 @@
 import { withCommander } from "react-native-salute";
 
-import environmentBanner from "../utils/environmentBanner";
-
-const withUtilities = (ScreenComponent) => {
-  const withEnvironmentBanner = environmentBanner(ScreenComponent);
-  const WithCommander = withCommander(withEnvironmentBanner);
-  return WithCommander;
-};
+const withUtilities = (ScreenComponent) => withCommander(ScreenComponent);
 
 export default withUtilities;
